pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "readme-ai-gen"
version = "1.0.0"
description = "Generate premium GitHub READMEs from any URL using AI"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "lekhan" }]
keywords = ["readme", "github", "ai", "cli", "generator", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "google-generativeai>=0.7",
    "openai>=1.30",
    "questionary>=2.0",
    "pyperclip>=1.8",
]

[project.scripts]
readme-gen = "readme_ai_gen.cli:main"

[project.urls]
Homepage = "https://github.com/lekhan/readme-ai-gen"
Repository = "https://github.com/lekhan/readme-ai-gen"
Issues = "https://github.com/lekhan/readme-ai-gen/issues"
